add ctor for containers

+++ main.cpp
@@ -2,11 +2,10 @@
 #include "bignumber.h"
 
 using namespace std;
-//using namespace BigInt;
 
 int main()
 {
-    BigInt a(8);
+    BigInt::BigInt a(8);
 
     a  / 4;
